APIs/azure.com/engagementfabric-EngagementFabric/2018-09-01-preview/python-aiohttp-client/tests/test_channels_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.channel import Channel
from openapi_server.models.channel_list import ChannelList
from openapi_server.models.cloud_error import CloudError


pytestmark = pytest.mark.asyncio

async def test_channels_create_or_update(client):
    """Test case for channels_create_or_update

    Create or Update the EngagementFabric channel
    """
    channel = {"properties":{"credentials":{"key":"credentials"},"channelFunctions":["channelFunctions","channelFunctions"],"channelType":"channelType"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.EngagementFabric/Accounts/{account_name}/Channels/{channel_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', channel_name='channel_name_example'),
        headers=headers,
        json=channel,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_channels_delete(client):
    """Test case for channels_delete

    Delete the EngagementFabric channel
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.EngagementFabric/Accounts/{account_name}/Channels/{channel_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', channel_name='channel_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_channels_get(client):
    """Test case for channels_get

    Get the EngagementFabric channel
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.EngagementFabric/Accounts/{account_name}/Channels/{channel_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example', channel_name='channel_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_channels_list_by_account(client):
    """Test case for channels_list_by_account

    List the EngagementFabric channels
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.EngagementFabric/Accounts/{account_name}/Channels'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', account_name='account_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

